radicle: check if Revision exists in Patch
If we have two operations, a = Reivision and b = Merge, then if we
have the following for any p = Patch then:

    p.apply([a, a, b]) =/= p.apply([a, b, a])

On the RHS we end up in the case that the reinserted Revision will
make the Revision redacted.

To stop this from happening, if a Revision is existing for the same
OpId then the `apply` is rejected.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett