cli: add semver option
If a consuming application wishes to obtain the semver using the
`--version` option, then it is required to parse the version output
for the semver itself.

This can be subject to change, so instead add a `--semver` option to
print only the semver version to stdout.

Signed-off-by: Fintan Halpenny <fintan.halpenny@gmail.com>
X-Clacks-Overhead: GNU Terry Pratchett