cli: allow to force push on a remote
The are cases where we need to perform a force push on a patch update,
so these commits allow passing this information from the command line.

A possible use case is when you rework the git history with some core
review and then you want to force push it.

At the current moment, the command fails, so change the code
to always force push should be simple enough to make the `rad update`
working most of the case.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>