cli: show a different identifier for the self user
Show a different identifier in the issue list is a good
the thing to do just to be able to sort see that there are issue
assigned to me.

It is also true that if I see the issue assigned to me
I can run the `rad issue list --assigned <did>` but I think
has a string `You` as we did in other places can be nice.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>