rad-patch: Fix patch open/upate when out-of-sync
Trying to open or update Patch will fail if the user's working copy if
out of sync with the Radicle projects canonical path.

Fix this by using an anonymous fetch to import the required commits
while avoiding affecting the user's working copy git references.  The
approach is already used for `rad-merge`.